feat: rename SimplePanel -> FlowPanel
feat: rename SimplePanel -> FlowPanel
